% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_level_shift.R
\name{add_level_shifts}
\alias{add_level_shifts}
\title{add_level_shifts}
\usage{
add_level_shifts(x, y)
}
\arguments{
\item{x}{datetimes}

\item{y}{shift vector}
}
\value{
vector of shifts
}
\description{
add_level_shifts
}
