% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{baro_kernel}
\alias{baro_kernel}
\title{baro_kernel}
\format{A \code{numeric vector}}
\usage{
baro_kernel
}
\description{
Example barometric response function kernel for testing.
}
\examples{
utils::data(baro_kern)
}
\keyword{datasets}
