% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_from_frf.R
\name{brf_from_frf}
\alias{brf_from_frf}
\title{brf_from_frf}
\usage{
brf_from_frf(x)
}
\arguments{
\item{x}{the complex transfer function vector}
}
\value{
vector of the barometric response function
}
\description{
Calculate the barometric response function from the transfer function
}
