% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frf_transfer_fun.R
\name{coh_phase}
\alias{coh_phase}
\title{coh_phase}
\usage{
coh_phase(pgram)
}
\arguments{
\item{pgram}{\code{numeric array} must be multivariate}
}
\value{
list of coherency and phase
}
\description{
calculate phase and coherency from the spectra and cross-spectra
}
