% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cos_harmonic}
\alias{cos_harmonic}
\title{cos_harmonic}
\usage{
cos_harmonic(x, freq)
}
\arguments{
\item{x}{\code{numeric vector} times to calculate sin and cos at}

\item{freq}{\code{numeric vector} frequencies for sin and cos}
}
\value{
cos curve
}
\description{
calculate sin and cos curves from POSIXct times (serial)
}
