% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributed_lag.R
\name{cross_basis_fft}
\alias{cross_basis_fft}
\title{cross_basis_fft}
\usage{
cross_basis_fft(basisvar, basislag)
}
\arguments{
\item{basisvar}{numeric vector}

\item{basislag}{lagging matrix}
}
\value{
numeric vector result of convolution
}
\description{
cross_basis_fft
}
