% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{det_parallel}
\alias{det_parallel}
\title{det_parallel}
\usage{
det_parallel(a)
}
\arguments{
\item{a}{\code{numeric array} values to evaluate}
}
\value{
vector of determinants
}
\description{
Determinant for an array in parallel
}
