% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_distributed_lag.R
\name{distributed_lag}
\alias{distributed_lag}
\title{distributed_lag}
\usage{
distributed_lag(x, knots, spline_fun = ns, lag_name = "",
  n_subset = 1, n_shift = 0)
}
\arguments{
\item{x}{numeric vector to lag}

\item{knots}{specific knots for the lagging process}

\item{spline_fun}{spline function to use i.e. splines::ns, splines::bs}

\item{lag_name}{name of variable to lag}

\item{n_subset}{for distributed_lag parallel}

\item{n_shift}{for distributed_lag parallel}
}
\value{
matrix with distributed lag terms
}
\description{
Depending on the vector to lag and the maximum knot, distributed_lag will 
either use an FFT (no NA and large maximum knot), or a parallel method 
(NA, or small maximum knot).
}
