% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distributed_lag_parallel}
\alias{distributed_lag_parallel}
\title{distributed_lag_parallel}
\usage{
distributed_lag_parallel(x, bl, lag_max)
}
\arguments{
\item{x}{matrix value of lag}

\item{bl}{matrix the basis lags}

\item{lag_max}{integer maximum number of lags}
}
\value{
distributed lag basis
}
\description{
This method calculates the basis for a distributed lag in parallel.  It is currently
slow.
}
