% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_energy_density.R
\name{energy_density_tsuboi}
\alias{energy_density_tsuboi}
\title{energy_density_tsuboi}
\usage{
energy_density_tsuboi(magnitude, distance)
}
\arguments{
\item{magnitude}{Richter magnitude}

\item{distance}{epicentral distance}
}
\description{
energy_density_tsuboi
}
\examples{
energy_density_tsuboi(5, 1000)
}
