% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_energy_density.R
\name{energy_density_wang}
\alias{energy_density_wang}
\title{energy_density_wang}
\usage{
energy_density_wang(magnitude, distance)
}
\arguments{
\item{magnitude}{Richter magnitude}

\item{distance}{epicentral distance}
}
\description{
energy_density_wang
}
\examples{
energy_density_wang(5, 9)
}
