% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_level_shift.R
\name{find_level_shift}
\alias{find_level_shift}
\title{find_level_shift}
\usage{
find_level_shift(x, dep_var = "val", time_var = "datetime",
  time_interval = 1)
}
\arguments{
\item{x}{the data set (data.table)}

\item{dep_var}{the dependent variable name (character)}

\item{time_var}{the time variable name (character)}

\item{time_interval}{the delta t (numeric)}
}
\value{
data.table of gaps
}
\description{
find_level_shift
}
