% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_level_shift.R
\name{fit_gaps}
\alias{fit_gaps}
\title{fit_gaps}
\usage{
fit_gaps(x, recipe, time_var = "datetime")
}
\arguments{
\item{x}{the data set (data.table)}

\item{recipe}{the recipe to apply}

\item{time_var}{the time variable name (character)}
}
\value{
data.table of fit results
}
\description{
fit_gaps
}
