% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_level_shift.R
\name{formula_from_recipe}
\alias{formula_from_recipe}
\title{formula_from_recipe
guess formula from recipe}
\usage{
formula_from_recipe(recipe)
}
\arguments{
\item{recipe}{recipe to apply}
}
\value{
formula
}
\description{
formula_from_recipe
guess formula from recipe
}
