% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frf_predict_terms.R
\name{frf_predict_terms}
\alias{frf_predict_terms}
\title{frf_predict_terms}
\usage{
frf_predict_terms(frf, x, time_col = "datetime", vars = c("wl",
  "baro"), coherency_cut = 0.3, limit = 2)
}
\arguments{
\item{frf}{frequency response function}

\item{x}{input variable dataset}

\item{time_col}{datetime}

\item{vars}{variables used in the transfer function call}

\item{coherency_cut}{values must have at least this coherency}

\item{limit}{gain must be lower than this value.  Often useful for Earth tides.}
}
\value{
data.table of results
}
\description{
frf_predict_terms
}
