% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_level_shift.R
\name{get_fit_summary}
\alias{get_fit_summary}
\title{get_fit_summary}
\usage{
get_fit_summary(x, recipe, start_reg, end_reg, start_gap, end_gap)
}
\arguments{
\item{x}{data.table of water levels}

\item{recipe}{recipe to apply}

\item{start_reg}{start subset time}

\item{end_reg}{end subset time}

\item{start_gap}{start subset time}

\item{end_gap}{end subset time}
}
\value{
data.table summary
}
\description{
get_fit_summary
}
