% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_level_shift.R
\name{get_level_shift_coef}
\alias{get_level_shift_coef}
\title{get_level_shift_coef}
\usage{
get_level_shift_coef(x)
}
\arguments{
\item{x}{data.table of level shifts from regression}
}
\value{
data.table of shifts
}
\description{
get_level_shift_coef
}
