% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_level_shift.R
\name{get_shift}
\alias{get_shift}
\title{get_shift}
\usage{
get_shift(x, recipe, start, end)
}
\arguments{
\item{x}{data.table of water levels}

\item{recipe}{recipe to apply}

\item{start}{start subset time}

\item{end}{end subset tiem}
}
\value{
data.table of predictions
}
\description{
get_shift
}
