% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{harmonic}
\alias{harmonic}
\title{harmonic}
\usage{
harmonic(times, freq)
}
\arguments{
\item{times}{\code{numeric vector} times to calculate sin and cos at}

\item{freq}{\code{numeric vector} frequencies for sin and cos}
}
\value{
sin and cos curves
}
\description{
calculate sin and cos curves from POSIXct times (serial)
}
