% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_pgram.R
\name{kern_apply}
\alias{kern_apply}
\title{kern_apply}
\usage{
kern_apply(x, k, circular = FALSE, ...)
}
\arguments{
\item{x}{\code{numeric vector} an input vector to be smoothed.}

\item{k}{\code{tskernel} smoothing "tskernel" object.}

\item{circular}{\code{logical} a logical indicating whether the input sequence to be smoothed is treated as circular, i.e., periodic.}

\item{...}{currently not used.}
}
\value{
smoothed periodogram
}
\description{
apply smoothing kernel using fftw
}
