% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{konno_ohmachi_parallel}
\alias{konno_ohmachi_parallel}
\title{konno_ohmachi_parallel}
\usage{
konno_ohmachi_parallel(x, f, b = 10L)
}
\arguments{
\item{x}{vector to smooth}

\item{f}{vector of frequencies}

\item{b}{integer even magnitude to smooth}
}
\value{
konno ohmachi smoothed results
}
\description{
This method does konno ohmachi smoothing
}
