% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{konno_ohmachi_serial}
\alias{konno_ohmachi_serial}
\title{konno_ohmachi_serial}
\usage{
konno_ohmachi_serial(x, f, b = 10L)
}
\arguments{
\item{x}{\code{numeric vector} to smooth}

\item{f}{\code{numeric vector} of frequencies}

\item{b}{\code{integer} smoothing coefficient}
}
\value{
vector of konno ohmachi smoothed results
}
\description{
This method does konno ohmachi smoothing
}
