% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_level_shift.R
\name{level_shift}
\alias{level_shift}
\title{level_shift}
\usage{
level_shift(x, level_breaks = NULL)
}
\arguments{
\item{x}{vector of numeric}

\item{level_breaks}{break points in the data}
}
\value{
vector of type factor
}
\description{
Create a factor column with the data subseted according to break points.
}
\examples{
level_shift(1:100, c(2, 40, 86))
}
