% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_level_shift.R
\name{level_shift_to_datetime}
\alias{level_shift_to_datetime}
\title{level_shift_to_datetime}
\usage{
level_shift_to_datetime(x)
}
\arguments{
\item{x}{character string that includes the date}
}
\value{
datetime
}
\description{
level_shift_to_datetime
}
