% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frf_spec_pgram.R
\name{pad_kernel}
\alias{pad_kernel}
\title{pad_kernel}
\usage{
pad_kernel(k, n)
}
\arguments{
\item{k}{\code{tskernel} smoothing "tskernel" object.}

\item{n}{\code{integer} the total length}
}
\value{
vector of the padded kernel
}
\description{
Pad a smoothing kernel
}
