% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portion.R
\name{portion}
\alias{portion}
\title{Extract Finalized Training Set}
\usage{
portion(object, ...)
}
\arguments{
\item{object}{A `recipe` object that has been prepared
with the option `retain = TRUE`.}

\item{...}{One or more selector functions to choose which variables will be
returned by the function. See \code{\link[=selections]{selections()}} for more details.
If no selectors are given, the default is to use
\code{\link[=everything]{everything()}}.}
}
\description{
As steps are estimated by `prep`, these operations are
 applied to the training set. Rather than running `bake`
 to duplicate this processing, this function will return
 variables from the processed training set.
}
\details{
When preparing a recipe, if the training data set is
 retained using `retain = TRUE`, there is no need to `bake` the
 recipe to get the preprocessed training set.

 `portion` will return the results of a recipes where _all steps_
 have been applied to the data, irrespective of the value of
 the step's `skip` argument.
}
\seealso{
[recipe()] [prep.recipe()] [bake.recipe()] [juice.recipe()]
}
