% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_level_shift.R
\name{predict_gaps}
\alias{predict_gaps}
\title{predict_gaps}
\usage{
predict_gaps(x, fits, term_labels = NULL)
}
\arguments{
\item{x}{the data set (data.table)}

\item{fits}{data.table of fits}

\item{term_labels}{names of the group (character)}
}
\value{
data.table of predictions
}
\description{
predict_gaps
}
