% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_terms.R
\name{predict_terms}
\alias{predict_terms}
\title{predict_terms}
\usage{
predict_terms(object, ...)
}
\arguments{
\item{object}{model fit from lm object}

\item{...}{currently not used}
}
\value{
return matrix of predictions
}
\description{
predict by term for 'mlm' or 'lm' object
}
