% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shift_subset}
\alias{shift_subset}
\title{shift_subset}
\usage{
shift_subset(x, lag, n_subset, n_shift)
}
\arguments{
\item{x}{\code{numeric vector} to lag}

\item{lag}{\code{integer vector} with the lags}

\item{n_subset}{\code{integer} subset every n_subset values}

\item{n_shift}{\code{integer} amount to shift results}
}
\value{
vector with lagged values
}
\description{
lag data and subset the results
}
