% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sin_harmonic}
\alias{sin_harmonic}
\title{sin_harmonic}
\usage{
sin_harmonic(x, freq)
}
\arguments{
\item{x}{\code{numeric vector} times to calculate sin and cos at}

\item{freq}{\code{numeric vector} frequencies for sin and cos}
}
\value{
sin curve
}
\description{
calculate sin and cos curves from POSIXct times (serial)
}
