% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slug_hvorslev.R
\name{slug_hvorslev}
\alias{slug_hvorslev}
\title{slug_hvorslev}
\usage{
slug_hvorslev(elapsed_time, hydraulic_conductivity, screen_length,
  radius_well, radius_filter)
}
\arguments{
\item{elapsed_time}{elapsed time (numeric)}

\item{hydraulic_conductivity}{formation hydraulic conductivity (numeric)}

\item{screen_length}{length of the screen (numeric)}

\item{radius_well}{well radius (numeric)}

\item{radius_filter}{radius of the filter pack (numeric)}
}
\value{
hvorslev slugtest drawdown
}
\description{
Response to a slug test
}
\references{
Hvorslev, M. J., Time lag and soil permeability in ground water
 measurements, Bull. 36, 50 pp., U.S. Army Corps of Eng. Waterways Exp. Stn.,
 Vicksburg, Miss., 1951
}
