% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_tf_parallel}
\alias{solve_tf_parallel}
\title{solve_tf_parallel}
\usage{
solve_tf_parallel(a)
}
\arguments{
\item{a}{\code{numeric array} values to evaluate}
}
\value{
vector of transfer functions
}
\description{
Calculate the transfer functions for an array in parallel
}
