% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frf_transfer_fun.R
\name{spec_welch_tf_error}
\alias{spec_welch_tf_error}
\title{spec_welch_tf_error}
\usage{
spec_welch_tf_error(coh, amp, overlap = 0.5, n_subsets = 10,
  return = "amp")
}
\arguments{
\item{coh}{coherency}

\item{amp}{amplitude}

\item{overlap}{overlap fraction from spec_welch}

\item{n_subsets}{number of subsets in spec welch}

\item{return}{either 'amp', or 'phase'}
}
\value{
error
}
\description{
spec_welch_tf_error
}
