% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_harmonic.R
\name{step_harmonic}
\alias{step_harmonic}
\title{Generate harmonics}
\usage{
step_harmonic(recipe, ..., role = "harmonic", trained = FALSE,
  freq = 1, prefix = "harmonic_", default = NA, columns = NULL,
  skip = FALSE, id = rand_id("harmonic"))
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose which
variables are affected by the step. See [selections()]
for more details. For the `tidy` method, these are not
currently used.}

\item{role}{Defaults to "harmonic"}

\item{trained}{A logical to indicate if the quantities for preprocessing
have been estimated.}

\item{freq}{\code{numeric vector} frequencies for sin and cos}

\item{prefix}{A prefix for generated column names, default to 
"harmonic_".}

\item{default}{Passed to \code{dplyr::lag}, determines what fills empty rows
left by lagging (defaults to NA).}

\item{columns}{A character string of variable names that will
be populated (eventually) by the \code{terms} argument.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake.recipe]{bake.recipe()}}? While all operations are baked
when \code{\link[=prep.recipe]{prep.recipe()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of `recipe` with the new step
 added to the sequence of existing steps (if any). For the
 `tidy` method, a tibble with columns `terms` which is
 the columns that will be affected and `holiday`.
}
\description{
`step_harmonic` creates a *specification* of a recipe step
 for harmonics (sin and cos terms).
}
\details{
`step_harmonic` calculates the Earth tide harmonics
}
\examples{
data(transducer)

rec <- recipe(wl ~ .,
              data = transducer[1:1000, list(datetime, wl)])

with_et <- rec \%>\%
  step_harmonic(datetime, freq = c(1, 1.93, 2)) \%>\% 
  prep() \%>\%
  juice()

}
\seealso{
[step_earthtide()] [recipe()]
  [prep.recipe()] [bake.recipe()]
}
\concept{generate harmonics of frequency in cycles per day}
\keyword{datagen}
