% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_lag_earthtide.R
\name{step_lag_earthtide}
\alias{step_lag_earthtide}
\title{Lagged Earth tide response}
\usage{
step_lag_earthtide(recipe, ..., role = "lag_earthtide",
  trained = FALSE, method = "gravity", lag = 0, astro_update = 1L,
  latitude = 0, longitude = 0, elevation = 0, azimuth = 0,
  gravity = 0, earth_radius = 6378136.3,
  earth_eccen = 0.0066943979514, cutoff = 1e-06, wave_groups = NULL,
  catalog = "ksm04", eop = NULL, prefix = "lag_earthtide_",
  default = NA, columns = NULL, skip = FALSE,
  id = rand_id("lag_earthtide"))
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose which
variables are affected by the step. See [selections()]
for more details. For the `tidy` method, these are not
currently used.}

\item{role}{Defaults to "lag_earthtide"}

\item{trained}{A logical to indicate if the quantities for preprocessing
have been estimated.}

\item{method}{One or more of "gravity", 
"tidal_potential", "tidal_tilt", "vertical_displacement", 
"horizontal_displacement", "n_s_displacement", "e_w_displacement",
"vertical_strain", "areal_strain", "volume_strain", or "ocean_tides",
"pole_tide", "lod_tide". The pole tide and lod_tide are used in predict
mode even if do_predict is FALSE. More than one value can only be used
if do_predict == TRUE.}

\item{lag}{A vector of positive integers. Each specified column will be
lagged for each value in the vector.}

\item{astro_update}{Integer that
determines how often to phases are updated in number of samples. Defaults
to 1 (every sample), but speed gains are realized with larger values.
Typically updating every hour will have speed gains and keep precision 
(ie 3600 for one second data, 60 for minute data, 1 for hourly data).}

\item{latitude}{The station latitude (numeric) defaults to 0.}

\item{longitude}{The station longitude (numeric) defaults to 0.}

\item{elevation}{The station elevation (m) (numeric) defaults to 0.}

\item{azimuth}{Earth azimuth (numeric) defaults to 0.}

\item{gravity}{Gravity at the station (m/s^2) (numeric) 0 to 
estimate gravity from elevation and latitude.}

\item{earth_radius}{Radius of earth (m) (numeric) defaults to 6378136.3}

\item{earth_eccen}{Eccentricity of earth (numeric) 
defaults to 6.69439795140e-3}

\item{cutoff}{Cutoff amplitude for constituents (numeric) 
defaults to 1e-6.}

\item{wave_groups}{Two column data.frame having start and end of 
frequency groups (data.frame). This data.frame must have two columns
with the names 'start', and 'end' signifying the start and end of the 
wave groupings.  An optional third column 'multiplier' can be provided 
to scale the particular wave group.  If column names do no match, the
inferred column positions are start, end, multiplier.}

\item{catalog}{Use the "hw95s" catalog or "ksm04" catalog (character).}

\item{eop}{User defined Earth Orientation Parameter (EOP) data.frame with the 
following columns: datetime, ddt, ut1_utc, lod, x, y, dx, dy}

\item{prefix}{A prefix for generated column names, default to 
"lag_earthtide_".}

\item{default}{Passed to \code{dplyr::lag}, determines what fills empty rows
left by lagging (defaults to NA).}

\item{columns}{A character string of variable names that will
be populated (eventually) by the \code{terms} argument.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake.recipe]{bake.recipe()}}? While all operations are baked
when \code{\link[=prep.recipe]{prep.recipe()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of `recipe` with the new step
 added to the sequence of existing steps (if any). For the
 `tidy` method, a tibble with columns `terms` which is
 the columns that will be affected and `holiday`.
}
\description{
`step_lag_earthtide` creates a *specification* of a recipe step
 that are lagged versions of the Earth tide response at a particular 
 location.  Requires the *earthtide* package
}
\details{
`step_lag_earthtide` calculates the earthtide response and then
 lags (leads) the terms.
}
\examples{
data(transducer)

rec <- recipe(wl ~ .,
              data = transducer[1:1000, list(datetime, wl)])

with_et <- rec \%>\%
  step_lag_earthtide(datetime, latitude = 34, longitude = -118.5, lag = -1:1) \%>\% 
  prep() \%>\%
  juice()

}
\seealso{
[step_lag_earthtide()] [recipe()]
  [prep.recipe()] [bake.recipe()]
}
\concept{generate a lagged form of the Earth tide response}
\keyword{datagen}
