% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_level_shift.R
\name{stretch_interp}
\alias{stretch_interp}
\title{stretch_interp}
\usage{
stretch_interp(start_val = NA, end_val = NA, values)
}
\arguments{
\item{start_val}{}

\item{end_val}{}

\item{values}{}
}
\value{
shifted vector of values
}
\description{
stretch_interp
}
