% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_summarize_lm.R
\name{summarize_coef}
\alias{summarize_coef}
\title{summarize_coef}
\usage{
summarize_coef(fit)
}
\arguments{
\item{fit}{model fit from lm object}
}
\value{
data.table with coefficient results
}
\description{
summarize_coef
}
