% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_summarize_lm.R
\name{summarize_lm.mlm}
\alias{summarize_lm.mlm}
\title{summarize_lm}
\usage{
\method{summarize_lm}{mlm}(fit)
}
\arguments{
\item{fit}{model fit from lm object}
}
\value{
data.table with fit results
}
\description{
summarize_lm
}
