% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_baro.R
\name{synthetic}
\alias{synthetic}
\title{synthetic}
\usage{
synthetic(sd_noise = 2e-04, sd_noise_trend = 3e-05,
  linear_trend = 1e-07, n = 14 * 86400, seed = NULL,
  baro_kernel = NULL)
}
\arguments{
\item{sd_noise}{standard deviation of random noise to add (numeric)}

\item{sd_noise_trend}{standard deviation of noise to add to generate a 
trend (numeric)}

\item{linear_trend}{magnitude of linear trend in time (numeric)}

\item{n}{length of time series in seconds (integer)}

\item{seed}{random number seed for reproducibility (numeric)}

\item{baro_kernel}{vector values to convolve with barometric pressure}
}
\value{
data.table of results
}
\description{
This function is used for testing purposes
}
