% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_baro.R
\name{synthetic_wl}
\alias{synthetic_wl}
\title{synthetic_wl}
\usage{
synthetic_wl(baro, datetime, sd_noise = 0, linear_trend = 0,
  intercept = 0, seed = NULL, kernel = NULL)
}
\arguments{
\item{baro}{barometric pressure (numeric vector)}

\item{datetime}{POSIXct dates}

\item{sd_noise}{standard deviation of random noise to add (numeric)}

\item{linear_trend}{magnitude of linear trend in time (numeric)}

\item{intercept}{}

\item{seed}{random number seed for reproducibility (numeric)}

\item{kernel}{}
}
\value{
data.table of synthetic water levels and barometric pressure
}
\description{
This function is used for testing purposes
}
