% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_fun.R
\name{transfer_fun}
\alias{transfer_fun}
\title{transfer_fun}
\usage{
transfer_fun(dat, vars, time = "datetime", method = "spec_pgram", ...)
}
\arguments{
\item{dat}{data that has the independent and dependent variables (data.table)}

\item{vars}{variables to include for transfer function calculation}

\item{time}{the name of the column that contains the POSIXct date and time}

\item{method}{either spec_pgram or spec_welch}

\item{...}{arguments to pass to method}
}
\value{
transfer function, gain and phase, coherency
}
\description{
transfer_fun
}
