% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vadose_response}
\alias{vadose_response}
\title{vadose_response}
\usage{
vadose_response(time, D, L, precision = 1e-08)
}
\arguments{
\item{time}{\code{numeric} value of elapsed time}

\item{D}{\code{numeric} unsaturated zone air diffusivity}

\item{L}{\code{numeric} unsaturated zone thickness}
}
\value{
brf from weeks 1979 model
}
\description{
weeks_1979 1-D air diffusivity
}
