% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributed_lag.R
\name{distributed_lag}
\alias{distributed_lag}
\title{distributed_lag}
\usage{
distributed_lag(x, knots, spline_fun, lag_name = "")
}
\arguments{
\item{x}{numeric vector to lag}

\item{knots}{specific knots for the lagging process}

\item{spline_fun}{spline function to use i.e. splines::ns, splines::bs}

\item{lag_name}{name of variable to lag}
}
\value{
matrix with distributed lag terms
}
\description{
distributed_lag
}
