% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/distributed_lag.R
\name{distributed_lag}
\alias{distributed_lag}
\title{distributed_lag}
\usage{
distributed_lag(x, knots, spline_fun = ns, lag_name = "")

distributed_lag(x, knots, spline_fun = ns, lag_name = "")
}
\arguments{
\item{x}{numeric vector to lag}

\item{knots}{specific knots for the lagging process}

\item{spline_fun}{spline function to use i.e. splines::ns, splines::bs}

\item{lag_name}{name of variable to lag}

\item{basisvar}{matrix value of lag}

\item{basislag}{matrix the basis lags}

\item{lag_max}{integer maximum number of lags}
}
\value{
distributed lag basis

matrix with distributed lag terms
}
\description{
This method calculates the basis for a distributed lag.  It is currently
slow.

Depending on the vector to lag and the maximum knot, distributed_lag will 
either use an FFT (no NA and large maximum knot), or a parallel method 
(NA, or small maximum knot).
}
