% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_level_shift.R
\name{get_fit_summary}
\alias{get_fit_summary}
\title{get_fit_summary}
\usage{
get_fit_summary(x, recipe, start, end)
}
\arguments{
\item{x}{data.table of water levels}

\item{recipe}{recipe to apply}

\item{start}{start subset time}

\item{end}{end subset tiem}
}
\value{
data.table of sumamry
}
\description{
get_fit_summary
}
