% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_level_shift.R
\name{get_intercept_stats}
\alias{get_intercept_stats}
\title{get_intercept_stats}
\usage{
get_intercept_stats(x)
}
\arguments{
\item{x}{data.table of level shifts from regression}
}
\value{
data.table of shift summaries
}
\description{
get_intercept_stats
}
