% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_response_from_fit.R
\name{response_from_fit}
\alias{response_from_fit}
\title{response_from_fit}
\usage{
response_from_fit(x, ...)
}
\arguments{
\item{x}{model fit object, commonly lm}

\item{...}{additional arguments}
}
\value{
tibble of model fit
}
\description{
response_from_fit
}
