% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frf_spec_from_pgram.R
\name{spec_from_pgram}
\alias{spec_from_pgram}
\title{spec_from_pgram}
\usage{
spec_from_pgram(pgram, method = "spec_pgram", ...)
}
\arguments{
\item{pgram}{the periodogram}

\item{...}{additional arguments}
}
\value{
spectrum from the pgram
}
\description{
spec_from_pgram
}
